Move serialize's tests
