+++ json/zio-json/src/test/scala/sttp/client4/ziojson/BackendStubZioJsonTests.scala
@@ -26,14 +26,4 @@ class BackendStubJson4sTests extends AnyFlatSpec with Matchers with ScalaFutures
     r.is200 should be(true)
     r.body should be(Right(Person("John")))
   }
-
-  it should "serialize from Json.Obj using implicit zioJsonBodySerializer" in {
-    val fields: Chunk[(String, Json)] = Chunk(("location", Json.Str("hometown")), ("bio", Json.Str("Scala programmer")))
-    val jObject: Json.Obj = Json.Obj(fields)
-    val result = basicRequest.get(Uri("http://example.org")).body(jObject).body.show
-
-    val expectedResult = "string: {\"location\":\"hometown\",\"bio\":\"Scala programmer\"}"
-
-    result should be(expectedResult)
-  }
 }
